Fix test failure of Environment.runningTime on Mac.
In the process, remove the terrible practice of settings global timer
in a library. Also, kill the static constructor. Now people don't need
to pay for timers unless they use them. Requires changing runningTime
to monotonicTime, as no code runs on startup to record the startup time.

Windows backend not tested and still needs work done to it.